Move method bodies out of header files
